+++ ocr.py
@@ -1,99 +1,80 @@
-from utils import get_vid
+import datetime
+import cv2
+import tempfile
+import json
 
-def prepare_ocr_visualization(mmif, view):
-    """ Visualize OCR by extracting image frames with BoundingBoxes from video"""
-    frames, text_docs, alignments = {}, {}, {}
-
-    for anno in view.annotations:
-        try:
-            # Save frames w/ BoundingBoxes to "frames" dict
-            if str(anno.at_type).endswith('BoundingBox'):
-                frame_num = anno.properties["frame"]
-                box_id = anno.properties["id"]
-                boxType = anno.properties["boxType"]
-                coordinates = anno.properties["coordinates"]
-                x = coordinates[0][0]
-                y = coordinates[0][1]
-                w = coordinates[3][0] - x
-                h = coordinates[3][1] - y
-                box = [box_id, boxType, [x, y, w, h]]
-
-                if frame_num in frames.keys():
-                    frames[frame_num]["boxes"].append(box)
-                    frames[frame_num]["bb_ids"].append(box_id)
-                else:
-                    frames[frame_num] = {"boxes": [box], "text": [], "bb_ids": [box_id], "timestamp": None, "secs": None, "repeat": False}
-
-            elif str(anno.at_type).endswith('TextDocument'):
-                t = anno.properties["text_value"]
-                if t:
-                    text_id = anno.properties["id"]
-                    # Format string so it is JSON-readable
-                    text_docs[text_id] = re.sub(r'[^\w]', '', t)
-
-            elif str(anno.at_type).endswith('Alignment'):
-                source = anno.properties["source"]
-                target = anno.properties["target"]
-                alignments[source] = target
-
-        except Exception as e:
-            print(f"Unexpected error of type {type(e)}: {e}")
-            pass
-
-    vid_path = get_video_path(mmif)
-    frames_list = [(k, v) for k, v in frames.items()]
-    return render_ocr(vid_path, frames_list, alignments, text_docs, 0, 1)
+from flask import render_template
 
 
-def render_ocr(vid_path, frames, alignments, text_docs, lower_n, upper_n):
+def add_bounding_box(anno, frames):
+    frame_num = anno.properties["frame"]
+    box_id = anno.properties["id"]
+    boxType = anno.properties["boxType"]
+    coordinates = anno.properties["coordinates"]
+    x = coordinates[0][0]
+    y = coordinates[0][1]
+    w = coordinates[3][0] - x
+    h = coordinates[3][1] - y
+    box = [box_id, boxType, [x, y, w, h]]
+    if frame_num in frames.keys():
+        frames[frame_num]["boxes"].append(box)
+        frames[frame_num]["bb_ids"].append(box_id)
+    else:
+        frames[frame_num] = {"boxes": [box], "text": [], "bb_ids": [box_id], "timestamp": None, "secs": None, "repeat": False}
+    return frames
+
+
+def align_annotations(frames_list, alignments, text_docs, fps):
+    """Link alignments with frames"""
+    prev_frame = None
+    for frame_num, frame in frames_list:
+        if fps:
+            secs = int(frame_num/fps)
+            frame["timestamp"] = str(datetime.timedelta(seconds=secs))
+            frame["secs"] = secs
+        for box_id in frame["bb_ids"]:
+            text_id = alignments[box_id]
+            frame["text"].append(text_docs[text_id])
+        if is_duplicate_ocr_frame(frame, prev_frame):
+            frame["repeat"] = True
+        prev_frame = frame
+    return frames_list
+
+
+def paginate(frames_list):
+    """Generate pages from a list of frames"""
+    pages = [[]]
+    n_frames_on_page = 0
+    for frame_num, frame in frames_list:
+        if n_frames_on_page >= 4 and not frame["repeat"]:
+            pages.append([])
+            n_frames_on_page = 0
+
+        pages[-1].append((frame_num, frame))
+
+        if not frame["repeat"]:
+            n_frames_on_page += 1
+
+    return pages
+
+def render_ocr(vid_path, frames_pages, page_number):
     """Iterate through frames and display the contents/alignments."""
-    # Read video as CV2 VideoCapture to extract screenshots + BoundingBoxes
-    cv2_vid = cv2.VideoCapture(vid_path)
-    fps = cv2_vid.get(cv2.CAP_PROP_FPS)
     # Path for storing temporary images generated by cv2
-    tmp_path = '/app/static/tmp'
-    if not os.path.exists(tmp_path):
-        os.makedirs(tmp_path)
-
-    prev_frame = {}
-    i, upper_i = 0, upper_n
-    while i <= upper_i:
-        if i >= len(frames) or i < 0:
-            break
-        frame_num, frame = frames[i]
-        # If frame has timestamp, it has already been processed
-        # if frame["timestamp"]:
-        #     break
+    cv2_vid = cv2.VideoCapture(vid_path)
+    for frame_num, frame in frames_pages[page_number]:
         cv2_vid.set(1, frame_num)
         _, frame_cap = cv2_vid.read()
         with tempfile.NamedTemporaryFile(
             prefix="/app/static/tmp/", suffix=".jpg", delete=False) as tf:
-            print(tf.name)
             cv2.imwrite(tf.name, frame_cap)
             # "id" is just the name of the temp image file
-            frame["id"] = tf.name[12:]
-            if fps:
-                secs = int(frame_num/fps)
-                frame["timestamp"] = datetime.timedelta(seconds=secs)
-                frame["secs"] = secs
-            for box_id in frame["bb_ids"]:
-                text_id = alignments[box_id]
-                frame["text"].append(text_docs[text_id])
-            
-            # Check for duplicates
-            if is_duplicate_ocr_frame(frame, prev_frame):
-                frame["repeat"] = True
-                upper_i += 1
-
-            prev_frame = frame
-            i += 1
+            frame["id"] = tf.name[12:]            
 
     return render_template('ocr.html', 
                            vid_path=vid_path, 
-                           frames=frames, 
-                           alignments=alignments, 
-                           text_docs=text_docs, 
-                           lower_n=lower_n, upper_n=upper_n)
+                           frames_pages=frames_pages, 
+                           page_number=page_number)
+
 
 def is_duplicate_ocr_frame(frame, prev_frame):
     if prev_frame:
