+++ utils.py
@@ -1,31 +1,17 @@
 import os
-import sys
-import datetime
-import secrets
-
 import cv2
-
-
 from io import StringIO
-from string import Template
-
 import displacy
-import requests
 import tempfile
 import re
-import ast
-import html
+import json
 
-from flask import Flask, request, render_template, flash, redirect, jsonify
-from werkzeug.utils import secure_filename
-
-from mmif.serialize import Mmif, View
+from flask import Flask, render_template
 from mmif.vocabulary import AnnotationTypes
-from mmif.vocabulary import DocumentTypes
 from lapps.discriminators import Uri
 from iiif_utils import generate_iiif_manifest
+from ocr import *
 
-FORWARD = 1
 
 # Get Properties from MMIF file ---
 
@@ -251,7 +237,20 @@ def url2posix(path):
 # Interactive MMIF Tab -----------
 
 def render_interactive_mmif(mmif):
+    # organize_alignments(mmif)
     return render_template('interactive.html', mmif=mmif)
+
+# def organize_alignments(mmif):
+#     """Organize all alignments so they are directly adjacent to the elements they align"""
+#     for view in mmif.views:
+#         if any([str(at_type).endswith('Alignment') for at_type in view.metadata.contains]):
+#             organize(view.annotations)
+
+# def organize(annotations):
+#     new_anno_list = []
+#     for annotation in annotations:
+#         new_anno_list.append(annotation)
+#     return new_anno_list
 
 # NER Tools ----------------------
 
@@ -288,7 +287,7 @@ def prepare_ocr_visualization(mmif, view):
                 if t:
                     text_id = anno.properties["id"]
                     # Format string so it is JSON-readable
-                    text_docs[text_id] = re.sub(r'[^\w]', '', t)
+                    text_docs[text_id] = re.sub(r'([\\\/\|\"\'])', r'\1 ', t)
 
             elif str(anno.at_type).endswith('Alignment'):
                 source = anno.properties["source"]
@@ -299,114 +298,11 @@ def prepare_ocr_visualization(mmif, view):
             print(f"Unexpected error of type {type(e)}: {e}")
             pass
 
+    # Generate pages (necessary to reduce IO cost) and render
     vid_path = get_video_path(mmif)
     cv2_vid = cv2.VideoCapture(vid_path)
     fps = cv2_vid.get(cv2.CAP_PROP_FPS)
     frames_list = [(k, v) for k, v in frames.items()]
     frames_list = align_annotations(frames_list, alignments, text_docs, fps)
     frames_pages = paginate(frames_list)
-    return render_ocr(vid_path, frames_pages, alignments, text_docs, 0)
-
-
-def add_bounding_box(anno, frames):
-    frame_num = anno.properties["frame"]
-    box_id = anno.properties["id"]
-    boxType = anno.properties["boxType"]
-    coordinates = anno.properties["coordinates"]
-    x = coordinates[0][0]
-    y = coordinates[0][1]
-    w = coordinates[3][0] - x
-    h = coordinates[3][1] - y
-    box = [box_id, boxType, [x, y, w, h]]
-    if frame_num in frames.keys():
-        frames[frame_num]["boxes"].append(box)
-        frames[frame_num]["bb_ids"].append(box_id)
-    else:
-        frames[frame_num] = {"boxes": [box], "text": [], "bb_ids": [box_id], "timestamp": None, "secs": None, "repeat": False}
-    return frames
-
-
-def align_annotations(frames_list, alignments, text_docs, fps):
-    prev_frame = None
-    for frame_num, frame in frames_list:
-        if fps:
-            secs = int(frame_num/fps)
-            frame["timestamp"] = datetime.timedelta(seconds=secs)
-            frame["secs"] = secs
-        for box_id in frame["bb_ids"]:
-            text_id = alignments[box_id]
-            frame["text"].append(text_docs[text_id])
-        if is_duplicate_ocr_frame(frame, prev_frame):
-            frame["repeat"] = True
-        prev_frame = frame
-    return frames_list
-
-def paginate(frames_list):
-    pages = [[]]
-    n_frames_on_page = 0
-    for frame_num, frame in frames_list:
-        if n_frames_on_page >= 4 and not frame["repeat"]:
-            pages.append([])
-            n_frames_on_page = 0
-
-        pages[-1].append((frame_num, frame))
-
-        if not frame["repeat"]:
-            n_frames_on_page += 1
-
-    return pages
-
-def render_ocr(vid_path, frames_pages, alignments, text_docs, page_number):
-    """Iterate through frames and display the contents/alignments."""
-    # Path for storing temporary images generated by cv2
-    cv2_vid = cv2.VideoCapture(vid_path)
-    for frame_num, frame in frames_pages[page_number]:
-        cv2_vid.set(1, frame_num)
-        _, frame_cap = cv2_vid.read()
-        with tempfile.NamedTemporaryFile(
-            prefix="/app/static/tmp/", suffix=".jpg", delete=False) as tf:
-            cv2.imwrite(tf.name, frame_cap)
-            # "id" is just the name of the temp image file
-            frame["id"] = tf.name[12:]            
-
-    return render_template('ocr.html', 
-                           vid_path=vid_path, 
-                           frames_pages=frames_pages, 
-                           alignments=alignments, 
-                           text_docs=text_docs, 
-                           page_number=page_number)
-
-
-def is_duplicate_ocr_frame(frame, prev_frame):
-    if prev_frame:
-        # Check Boundingbox distances
-        rounded_prev = round_boxes(prev_frame["boxes"])
-        for box in round_boxes(frame["boxes"]):
-            if box in rounded_prev and frame["secs"]-prev_frame["secs"] < 5:
-                return True
-    return False
-
-def round_boxes(boxes):
-    # To account for jittery bounding boxes in OCR annotations
-    rounded_boxes = []
-    for box in boxes:
-        rounded_box = []
-        for coord in box[2]:
-            rounded_box.append(round(coord/10)*10)
-        rounded_boxes.append(rounded_box)
-    return rounded_boxes
-
-def get_ocr_views(mmif):
-    """Return OCR views, which have TextDocument and Alignment annotations, but no
-    other annotations."""
-    views = []
-    # TODO: not sure why we use the full URL
-    needed_types = set([
-        "http://mmif.clams.ai/0.4.0/vocabulary/TextDocument",
-        "http://mmif.clams.ai/0.4.0/vocabulary/BoundingBox",
-        "http://mmif.clams.ai/0.4.0/vocabulary/Alignment" ])
-    for view in mmif.views:
-        annotation_types = view.metadata.contains.keys()
-        if needed_types.issubset(annotation_types) and len(annotation_types) == 3:
-            views.append(view)
-    return views
+    return render_ocr(vid_path, frames_pages, 0)
